+++ src/test/java/org/mps/Fibonacci/FibonacciTest.java
@@ -34,56 +34,56 @@ public class FibonacciTest {
 
     @Test
     public void ShouldReturnZeroIfNumberIsZero(){
-        int expectedValue = 0;
-        int obtainedValue = fib.compute(0);
+        double expectedValue = 0;
+        double obtainedValue = fib.compute(0);
 
         assertEquals(expectedValue, obtainedValue);
     }
 
     @Test
     public void ShouldReturnOneIfNumberIsOne(){
-        int expectedValue = 0;
-        int obtainedValue = fib.compute(0);
+        double expectedValue = 0;
+        double obtainedValue = fib.compute(0);
 
         assertEquals(expectedValue, obtainedValue);
     }
 
     @Test
     public void ShouldReturnOneIfNumberIsTwo(){
-        int expectedValue = 0;
-        int obtainedValue = fib.compute(0);
+        double expectedValue = 0;
+        double obtainedValue = fib.compute(0);
 
         assertEquals(expectedValue, obtainedValue);
     }
 
     @Test
     public void ShouldReturnTwoIfNumberIsThree(){
-        int expectedValue = 0;
-        int obtainedValue = fib.compute(0);
+        double expectedValue = 0;
+        double obtainedValue = fib.compute(0);
 
         assertEquals(expectedValue, obtainedValue);
     }
 
     @Test
     public void ShouldReturnThreeIfNumberIsFour(){
-        int expectedValue = 0;
-        int obtainedValue = fib.compute(0);
+        double expectedValue = 0;
+        double obtainedValue = fib.compute(0);
 
         assertEquals(expectedValue, obtainedValue);
     }
 
     @Test
     public void ShouldReturnFiveIfNumberIsFive(){
-        int expectedValue = 0;
-        int obtainedValue = fib.compute(0);
+        double expectedValue = 0;
+        double obtainedValue = fib.compute(0);
 
         assertEquals(expectedValue, obtainedValue);
     }
 
     @Test
     public void ShouldReturn6765IfNumberIs20(){
-        int expectedValue = 0;
-        int obtainedValue = fib.compute(0);
+        double expectedValue = 0;
+        double obtainedValue = fib.compute(0);
 
         assertEquals(expectedValue, obtainedValue);
     }
